+++ src/main/java/com/ludo/study/studymatchingplatform/auth/google/service/GoogleLoginService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 
 import com.ludo.study.studymatchingplatform.auth.google.GoogleOAuthToken;
 import com.ludo.study.studymatchingplatform.auth.google.GoogleUserInfo;
-import com.ludo.study.studymatchingplatform.user.domain.User;
-import com.ludo.study.studymatchingplatform.user.repository.UserRepositoryImpl;
+import com.ludo.study.studymatchingplatform.user.domain.user.User;
+import com.ludo.study.studymatchingplatform.user.repository.user.UserRepositoryImpl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
