+++ src/main/java/com/ludo/study/studymatchingplatform/auth/google/service/GoogleSignUpService.java
@@ -5,8 +5,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.ludo.study.studymatchingplatform.auth.google.GoogleOAuthToken;
 import com.ludo.study.studymatchingplatform.auth.google.GoogleUserInfo;
-import com.ludo.study.studymatchingplatform.user.domain.User;
-import com.ludo.study.studymatchingplatform.user.repository.UserRepositoryImpl;
+import com.ludo.study.studymatchingplatform.user.domain.user.User;
+import com.ludo.study.studymatchingplatform.user.repository.user.UserRepositoryImpl;
 
 import lombok.RequiredArgsConstructor;
