+++ src/main/java/com/ludo/study/studymatchingplatform/auth/kakao/service/KakaoOAuthTokenRequestService.java
@@ -11,7 +11,7 @@ import org.springframework.web.client.RestTemplate;
 
 import com.ludo.study.studymatchingplatform.auth.kakao.service.dto.KakaoOAuthToken;
 import com.ludo.study.studymatchingplatform.auth.naver.repository.InMemoryClientRegistrationAndProviderRepository;
-import com.ludo.study.studymatchingplatform.user.domain.Social;
+import com.ludo.study.studymatchingplatform.user.domain.user.Social;
 
 import lombok.RequiredArgsConstructor;
