+++ src/main/java/com/ludo/study/studymatchingplatform/auth/naver/service/NaverLoginService.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Service;
 import com.ludo.study.studymatchingplatform.auth.naver.service.dto.response.LoginResponse;
 import com.ludo.study.studymatchingplatform.auth.naver.service.vo.response.NaverOAuthToken;
 import com.ludo.study.studymatchingplatform.auth.naver.service.vo.response.UserProfile;
-import com.ludo.study.studymatchingplatform.user.domain.User;
-import com.ludo.study.studymatchingplatform.user.repository.UserRepositoryImpl;
+import com.ludo.study.studymatchingplatform.user.domain.user.User;
+import com.ludo.study.studymatchingplatform.user.repository.user.UserRepositoryImpl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
